Adicionando validacao para nao excluir uma conta com transacao associada

+++ src/services/account.js
@@ -25,7 +25,13 @@ module.exports = (app) => {
       .update(account, '*')
   }
 
-  const remove = (id) => {
+  const remove = async (id) => {
+    const transaction = await app.services.transaction.findOne({ acc_id: id })
+
+    if (transaction) {
+      throw new ValidationError('Essa conta possui transacoes associadas')
+    }
+
     return app.db('accounts')
       .where({ id })
       .delete()
